requirements.txt
Django==1.11
pytz==2019.1
RPi.GPIO==0.6.5
